Did a little styling on errors

+++ database/migrations/2016_06_02_183654_create_holidays_table.php
@@ -19,7 +19,7 @@ class CreateHolidaysTable extends Migration
 			$table->boolean('prebooked')->nullable();
 			$table->dateTime('request_date_from');
 			$table->dateTime('request_date_to');
-			$table->integer('saturday')->unsigned()->nullable();
+			$table->float('saturday')->unsigned()->nullable();
 			$table->boolean('approved')->nullable();
 			$table->timestamps();
         });
